test: use double precision in wrapped signature for old f2py
Older f2py (NumPy on Python 3.8) maps the iso_c_binding c_double kind to C
float, silently corrupting results. Declare the f2py-exposed dummy args as
plain double precision; keep bind(C) only on the internal C interface.

+++ tests/packages/cmix/mixed.f90
@@ -2,10 +2,11 @@
 ! Without the original source compiled into the module (or the C library
 ! linked), importing this fails with an undefined symbol (issue #20).
 subroutine add_them(a, b, c)
-  use iso_c_binding, only: c_double
   implicit none
-  real(c_double), intent(in) :: a, b
-  real(c_double), intent(out) :: c
+  ! Plain double precision in the wrapped signature: older f2py maps the
+  ! iso_c_binding c_double kind to C float, which silently corrupts results.
+  double precision, intent(in) :: a, b
+  double precision, intent(out) :: c
   interface
     function add_in_c(x, y) bind(c, name="add_in_c") result(r)
       use iso_c_binding, only: c_double
